backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
google-generativeai>=0.4.1
python-dotenv>=1.0.0